refactor: move resource and backup benchmark test to its dedicated directory
