+++ tests/bench/resource/resource_repo_test.go
@@ -1,4 +1,4 @@
-package bench
+package resource
 
 import (
 	"context"
